fix(guard): keep disconnect confirmation after a dry run
Leave leftover hooks fail-closed unless an app is explicitly disabled,
and keep the disconnect confirm step armed after a dry-run uninstall.
